requirements.txt
requests
azure-cosmos
azure-storage-blob
azure-identity
azure-mgmt-storage
azure.mgmt.resource
python-dotenv